refactor test fixtures and imports in conftest.py and test_tasks.py
